Added new Api calls and Security Update 2

+++ VistaExamenPlanner/Controllers/Klas.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MySql.Data.MySqlClient;
+using System.Security.Claims;
 using VistaExamenPlanner.Handler;
 using VistaExamenPlanner.Objecten;
 
@@ -8,7 +9,14 @@ namespace VistaExamenPlanner.Controllers
 {
     public class Klas : Controller
     {
-        [Authorize]
+
+        private readonly ILogger<Klas> _logger;
+
+        public Klas(ILogger<Klas> logger)
+        {
+            _logger = logger;
+        }
+
         [HttpGet]
         [Route("SelectAllKlasses")]
         public string SelectAllKlasses()
@@ -17,7 +25,7 @@ namespace VistaExamenPlanner.Controllers
             {
                 MySqlCommand SelectAllKlasses = new MySqlCommand();
 
-                SelectAllKlasses.CommandText = "SELECT * FROM Klas;";
+                SelectAllKlasses.CommandText = "SELECT Id, Naam FROM class;";
 
                 return database.Select(SelectAllKlasses);
             }
@@ -28,6 +36,13 @@ namespace VistaExamenPlanner.Controllers
         [Route("SelectOneKlasFromName")]
         public string SelectOneKlas([FromBody] string naam)
         {
+            string Rol = User.FindFirstValue("Rol");
+            if (Rol == "" || Rol != "3")
+            {
+                _logger.Log(LogLevel.Information, $"SelectOneKlasFromName: Someone without the rights tried to select a class from the name.");
+                return "403";
+            }
+
             using (DatabaseHandler database = new DatabaseHandler())
             {
                 MySqlCommand SelectOneKlas = new MySqlCommand();
@@ -42,8 +57,15 @@ namespace VistaExamenPlanner.Controllers
         [Authorize]
         [HttpPost]
         [Route("CreateKlas")]
-        public void CreateKlas(CreateKlass klas)
+        public void CreateKlas([FromBody] CreateKlass klas)
         {
+            string Rol = User.FindFirstValue("Rol");
+            if (Rol == "" || Rol != "3")
+            {
+                _logger.Log(LogLevel.Information, $"CreateKlas: Someone without the rights tried to create a class.");
+                return;
+            }
+
             using (DatabaseHandler database = new DatabaseHandler())
             {
                 MySqlCommand AddKlas = new MySqlCommand();
@@ -61,11 +83,18 @@ namespace VistaExamenPlanner.Controllers
         [Route("DeleteKlas")]
         public void DeleteKlas([FromBody] int id)
         {
+            string Rol = User.FindFirstValue("Rol");
+            if (Rol == "" || Rol != "3")
+            {
+                _logger.Log(LogLevel.Information, $"DeleteKlas: Someone without the rights tried to create a class.");
+                return;
+            }
+
             using (DatabaseHandler database = new DatabaseHandler())
             {
                 MySqlCommand DeleteKlas = new MySqlCommand();
 
-                DeleteKlas.CommandText = "DELETE FROM Student WHERE Klas_Id=@Id; DELETE FROM AgendaItem WHERE Klas_Id=@Id; DELETE FROM Klas WHERE Id=@Id;";
+                DeleteKlas.CommandText = "DELETE FROM User WHERE Klas_Id=@Id; DELETE FROM AgendaItem WHERE Klas_Id=@Id; DELETE FROM Klas WHERE Id=@Id;";
                 DeleteKlas.Parameters.AddWithValue("@Id", id);
 
                 database.Delete(DeleteKlas);
